Refs #25160 - Cleanup
- move launchers to FAC::TaskLauncher module
- rename group_runner_{class,input} to runner_{class,input}
