+++ lib/foreman_ansible_core.rb
@@ -22,13 +22,13 @@ module ForemanAnsibleCore
 
   require 'foreman_remote_execution_core/actions'
   require 'foreman_ansible_core/remote_execution_core/ansible_runner'
-  require 'foreman_ansible_core/playbook_task_launcher'
-  require 'foreman_ansible_core/ansible_runner_task_launcher'
+  require 'foreman_ansible_core/task_launcher/playbook'
+  require 'foreman_ansible_core/task_launcher/ansible_runner'
 
   if defined?(SmartProxyDynflowCore)
     SmartProxyDynflowCore::TaskLauncherRegistry.register('ansible-runner',
-                                                         ForemanAnsibleCore::AnsibleRunnerTaskLauncher)
+                                                         ForemanAnsibleCore::TaskLauncher::AnsibleRunner)
     SmartProxyDynflowCore::TaskLauncherRegistry.register('ansible-playbook',
-                                                         ForemanAnsibleCore::PlaybookTaskLauncher)
+                                                         ForemanAnsibleCore::TaskLauncher::Playbook)
   end
 end
